sync berry crush with emerald
